refactoring: after review + fixed routing forum error
