+++ packages/client/src/components/Layout/index.tsx
@@ -1,7 +1,7 @@
 import { ReactElement, useMemo } from 'react'
 import cn from './style.module.scss'
 import Header from '../UI/Header'
-import Alert from '../Alert'
+import Alert from '../UI/Alert'
 import Loader from '../Loader'
 import { useLocation } from 'react-router-dom'
 import { routerList } from '../RouterList/settings'
@@ -14,16 +14,18 @@ const Layout = ({ children }: Props) => {
   const { pathname } = useLocation()
 
   const { type, header } = useMemo(() => {
-    return routerList[pathname as keyof typeof routerList]
+    if (Object.hasOwn(routerList, pathname)) {
+      return routerList[pathname as keyof typeof routerList]
+    }
+    return { type: undefined, header: undefined }
   }, [pathname])
 
   if (type === 'Landing') {
     return <div className={cn.Layout}>{children}</div>
   }
-  const background = cn[type]
 
   return (
-    <div className={`${cn.Layout} ${background}`}>
+    <div className={`${cn.Layout} ${cn[type ?? 'Default']}`}>
       <div className={`${cn.Container}`}>
         <div className={cn.Wrapper}>
           <Header>{header}</Header>
