+++ packages/client/src/components/UI/Alert/index.tsx
@@ -1,15 +1,15 @@
 import { useSelector } from 'react-redux'
-import { IRootStore } from '../../store/reduces/interfaces'
+import { IRootStore } from '../../../store/reduces/interfaces'
 import cn from './style.module.scss'
 
 const Alert = () => {
   const { show, type, text } = useSelector((state: IRootStore) => state.alert)
 
-  if (show) {
-    return <div className={`${cn.Alert} ${cn[type]}`}>{text}</div>
+  if (!show) {
+    return null
   }
 
-  return null
+  return <div className={`${cn.Alert} ${cn[type]}`}>{text}</div>
 }
 
 export default Alert
