+++ packages/client/src/views/Forum/CreateTheme/style.module.scss
@@ -1,4 +1,4 @@
-@import 'src/styles/vars';
+@import './src/styles/vars';
 @import 'src/styles/mixins';
 
 .FormSendMsg {
